requirements.txt
asgiref==3.7.2
Django==5.0.3
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
Markdown==3.6
mysqlclient==2.2.4
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
